+++ utils/apply-apache-2-license.sh
@@ -1,3 +1,17 @@
+#!/usr/bin/env bash
+SCRIPT_ROOT=$(cd $(dirname $0); pwd)
+cd $SCRIPT_ROOT/..
+
+for i in $(find src -type f -name "*.ts")
+do
+    if grep -q "Licensed under the Apache License, Version 2.0" $i; then
+        echo "License already applied to $i"
+        continue;
+    fi
+
+    echo "Applying Apache 2.0 license to $i"
+
+    cat >> tmp.txt <<EOL
 /**
  * Copyright 2025 Angus.Fenying <fenying@litert.org>
  *
@@ -13,26 +27,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+EOL
 
-import * as ConfigLoader from '../../lib';
-import * as BuiltInEnc from '../../lib/Encodings';
-import { LocalFileReader } from '../../lib/Readers/LocalFileReader';
-import { UppercaseOperator } from './uppercase-operator';
+    echo "" >> tmp.txt
 
-const loader = new ConfigLoader.ConfigLoader({
-    reader: new LocalFileReader({
-        encodings: {
-            '.json': 'json',
-        }
-    }),
-    encodings: BuiltInEnc.getAllBuiltInEncodings(),
-    operators: [ new UppercaseOperator() ]
-});
+    cat $i >> tmp.txt
 
-(async () => {
+    mv tmp.txt $i
 
-    const ret = await loader.load(`${__dirname}/../../test-data/02/main.json`);
-
-    console.log(ret);
-
-})();
+done